Add screenPlacement support for Team Arena menus
Allow custom Team Arena HUDs to specify screen placement for
each menudef. This allows parts to be attached to edge of screen
in widescreen as already done by CGame for voice chat head and
vertical power up area.

It only affects the draw location. It's mainly for HUDs, not
interactive menus. Though it could be used for menu decorations.

Commit imported from mint-arena.